+++ libs/packages/components/src/lib/autocomplete-search/autocomplete-search.component.scss
