+++ libs/packages/components/src/lib/autocomplete-search/models/SDSAutocompleteConfiguration.ts
