+++ libs/packages/components/src/lib/selected-result/models/SDSSelectedResultConfiguration.ts
