+++ libs/packages/components/src/lib/selected-result/models/sds-selectedItem.model.ts
