+++ libs/packages/components/src/lib/selected-result/selected-result.component.html
